+++ ckanext/unhcr/migration/unhcr/versions/004_65d2d180530c_add_geography_models.py
@@ -7,7 +7,6 @@ Create Date: 2021-07-01 11:31:04.118909
 """
 from alembic import op
 import sqlalchemy as sa
-from ckanext.unhcr.models import Geography
 
 
 # revision identifiers, used by Alembic.
@@ -18,8 +17,17 @@ depends_on = None
 
 
 def upgrade():
-    Geography.__table__.create()
+    op.create_table(
+        'geography',
+        sa.Column('globalid', sa.UnicodeText, primary_key=True),
+        sa.Column('pcode', sa.UnicodeText, nullable=False),
+        sa.Column('iso3', sa.UnicodeText, nullable=False),
+        sa.Column('gis_name', sa.UnicodeText, nullable=False),
+        sa.Column('gis_status', sa.Enum('active', 'inactive', name='geography_gis_statuse_enum'), nullable=False),
+        sa.Column('layer', sa.UnicodeText, nullable=False),
+        sa.Column('hierarchy_pcode', sa.UnicodeText, nullable=False)
+    )
 
 
 def downgrade():
-    Geography.__table__.drop()
+    op.drop_table('geography')