+++ ckanext/unhcr/src/js/activities-list.js
@@ -1,0 +1,11 @@
+$( document ).ready(function() {
+
+	$("#system-activities-list").fancyTable({
+		pagination: true,
+		inputPlaceholder: "Filter activities list table ...",
+		perPage:20,
+		globalSearch:true,
+		inputStyle: "width: inherit"
+	});
+
+});