+++ CHANGES.md
@@ -1,6 +1,8 @@
 # RIDL Changelog
 
 ## v3.1.7 - 2021-XX-XX
+New features:
+ - Tracking system activities [#734](https://github.com/okfn/ckanext-unhcr/pull/734)
 Bug fixes:
  - Show deposit errors [#733](https://github.com/okfn/ckanext-unhcr/pull/733)